+++ src/main/java/com/mydigitalschool/dao_orm/daojooq/business/Components.java
@@ -3,7 +3,7 @@ package com.mydigitalschool.dao_orm.daojooq.business;
 import com.mydigitalschool.dao_orm.daojooq.dao.tables.ParcAttractions;
 
 public class Components { //tarif_parc
-    public long id;
-    public long idParc;
+    public Integer id;
+    public Integer idParc;
     public float tarif;
 }
